requirements.txt
# Python 3.7.9 compatible stack (Django 4.x requires Python 3.8+)
Django>=3.2.25,<4.0
djangorestframework>=3.14.0,<3.15
djangorestframework-simplejwt>=5.2.0,<5.3
channels>=3.0.5,<4.0
daphne>=3.0.2,<4.0
PyYAML>=6.0.1,<7.0
paramiko>=3.3.0,<4.0
requests>=2.31.0,<3.0
django-cors-headers>=3.13.0,<4.0
cryptography>=41.0.0,<43.0
